facebook-scraper/requirements.txt
bs4
selenium
python-dateutil
lxml